WIP Getenv returns ContextVar (bad idea)
For DEBUG ContextVar makes sense since it is only used as integer and
for comparisons to make print statements.

But here one might expect a environment variable of type string and
might want to operate on it directly which is not possible.

For example:

  getenv('CL_EXCLUDE', "").value.split(",")

So I will scrap this idea I think. The upside is also not clear as it
was mostly just a hunch I wante to try out.